BANGS HEAD.... need to generate boot.ext4 before flashing...

+++ 07_fastboot_emmc.sh
@@ -6,10 +6,18 @@ if ! id | grep -q root; then
 	exit
 fi
 
-fastboot flash ram ./deploy/u-boot-with-spl.bin
-fastboot reboot
-sleep 10
-fastboot flash uboot ./deploy/u-boot-with-spl.bin
-fastboot flash boot ./deploy/boot.ext4
-fastboot flash root ./deploy/root.ext4
-fastboot reboot
+if [ -f ./.05_generate_boot.sh ] ; then
+	echo "ERROR: run: [sudo ./05_generate_boot.sh] first"
+else
+	if [ -f ./.06_generate_root.sh ] ; then
+		echo "ERROR: run: [sudo ./06_generate_root.sh] first"
+	else
+		fastboot flash ram ./deploy/u-boot-with-spl.bin
+		fastboot reboot
+		sleep 10
+		fastboot flash uboot ./deploy/u-boot-with-spl.bin
+		fastboot flash boot ./deploy/boot.ext4
+		fastboot flash root ./deploy/root.ext4
+		fastboot reboot
+	fi
+fi
